requirements.txt
deepspeed==0.9.4
gradio==3.36.1
librosa==0.10.0
moviepy==1.0.3
numpy==1.25.0
openai_whisper==20230314
opencv_python==4.7.0.72
orjson==3.9.2
Pillow==9.0.1
Pillow==10.0.0
pydub==0.25.1
Requests==2.31.0
scipy==1.10.1
setuptools==67.8.0
soundfile==0.12.1
tiktoken==0.3.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.30.2
